requirements.txt
fastapi
uvicorn
python-multipart
pydantic
PyPDF2
sentence-transformers
transformers
torch
numpy
scikit-learn
python-dotenv
